Refactor out comma whitespace checks into CommaWhitespaceListener
