+++ src/test/java/com/sleekbyte/tailor/functional/ArrowWhitespaceTest.java
@@ -4,7 +4,10 @@ import com.sleekbyte.tailor.common.Messages;
 import com.sleekbyte.tailor.common.Rules;
 import com.sleekbyte.tailor.common.Severity;
 import com.sleekbyte.tailor.output.Printer;
+import org.junit.runner.RunWith;
+import org.mockito.runners.MockitoJUnitRunner;
 
+@RunWith(MockitoJUnitRunner.class)
 public class ArrowWhitespaceTest extends RuleTest {
 
     @Override
